Added emergency SD load on stage2
If an exception is triggered, an emergency load from SD happens. The
payload is loaded from the very beginning of the SD card (Sector 0), not
from the filesystem in the card. The emergency loading has not been
tested, but at least adding the code in did not cause an o3DS to fail to
load anything else.